Save the combined and the final graph representing dams in Myanmar
Combined graph is a graph of connections between dams in all water
networks combined, i.e. all basins share the same outlet node. In this
graph dams are nodes and edges represent the connections (river sections)
between neighbouring dams.

The final graph is the graph used for generating input file to the MOO
algorithm that solves value problem on graph structured networks. In the
final graph the network is transformed such that the edges represent dams
and the nodes represents river sections between dams. The nodes and the
edges are renamed to consecutive integers between 0 and 1 in order to
conform to the data requirements imposed by the MOO solver.