feat: Add comparison of explicit and simplified emissions
Provides a new jupyter notebook that compares emissions calculated with ReEmission
with the simplified per climate zone estimates of Soued et al. 2022. The calculates
emissions for all reservoirs in Myanmar using the emission factors of Soued et al.
and compares the differences between the emission estimates from emission factors
and from ReEmission. The final dataframe with all calculated values is saved to
an output csv and excel file.

Additionally, main README.md is updated with the short description of the
notebook.